cloud-run-test/requirements.txt
Flask==2.0.1
requests==2.26.0
ptvsd==4.3.2 # Required for debugging.
Pillow==8.3.1
numpy
opencv-python
-f https://dl.fbaipublicfiles.com/detectron2/wheels/cpu/index.html
detectron2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu
torchaudio==0.9.0
google-cloud-storage
